Move helper functions to helpers.py and apply final UI fixes
