+++ templates/security/cves/about.html
@@ -305,13 +305,13 @@
               </tr>
               <tr>
                 <td class="p-table__cell--icon-placeholder">
-                  <i class="p-icon--warning"></i>Vulnera&shy;ble
+                  <i class="p-icon--error"></i>Vulnera&shy;ble
                 </td>
                 <td colspan="2">The package is affected by the issue for the given release.</td>
               </tr>
               <tr>
                 <td class="p-table__cell--icon-placeholder">
-                  <i class="p-icon--warning"></i>Vulnera&shy;ble, work in progress
+                  <i class="p-icon--error"></i>Vulnera&shy;ble, work in progress
                 </td>
                 <td colspan="2">
                   The package is affected by the issue for the given release, and the Ubuntu security team is currently working on a fix.
@@ -319,7 +319,7 @@
               </tr>
               <tr>
                 <td class="p-table__cell--icon-placeholder">
-                  <i class="p-icon--warning"></i>Vulnera&shy;ble, fix deferred
+                  <i class="p-icon--error"></i>Vulnera&shy;ble, fix deferred
                 </td>
                 <td colspan="2">
                   The package is affected by the issue for the given release, but fixing has been deferred. We usually provide an explanation, either next to the status or in the notes.
